Revert "moved AprilTags to its own class and refactored the parking code"
