+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/AprilTagParking3231.java
@@ -38,7 +38,6 @@ import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.robot.Robot;
 
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
-import org.firstinspires.ftc.teamcode.lib.AprilTagsCamera;
 import org.firstinspires.ftc.teamcode.lib.Grabber;
 import org.firstinspires.ftc.teamcode.lib.LinearSlider;
 import org.firstinspires.ftc.teamcode.lib.RobotCentricMechanumDrive;
@@ -53,53 +52,163 @@ import java.util.ArrayList;
 @Autonomous(name= "Custom Sleeve Parking (3231)", group="Linear Opmode")
 public class AprilTagParking3231 extends LinearOpMode {
 
+    // TODO: EDIT after measuring
     static final double TICKS_PER_INCH = 50;
     static final double TICKS_PER_DEGREE = 50;
 
+    OpenCvCamera camera;
+    AprilTagDetectionPipeline aprilTagDetectionPipeline;
+
+    static final double FEET_PER_METER = 3.28084;
+
+    // Lens intrinsics
+    // UNITS ARE PIXELS
+    // NOTE: this calibration is for the C920 webcam at 800x448.
+    // You will need to do your own calibration for other configurations!
+    double fx = 578.272;
+    double fy = 578.272;
+    double cx = 402.145;
+    double cy = 221.506;
+
+    // UNITS ARE METERS
+    double tagsize = 0.166;
+
+    int LEFT = 11;
+    int MIDDLE = 12;
+    int RIGHT = 13;
+
+    AprilTagDetection tagOfInterest = null;
+
     @Override
     public void runOpMode() throws InterruptedException {
+        int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier("cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
+        camera = OpenCvCameraFactory.getInstance().createWebcam(hardwareMap.get(WebcamName.class, "Webcam 1"), cameraMonitorViewId);
+        aprilTagDetectionPipeline = new AprilTagDetectionPipeline(tagsize, fx, fy, cx, cy);
+
+        camera.setPipeline(aprilTagDetectionPipeline);
+        camera.openCameraDeviceAsync(new OpenCvCamera.AsyncCameraOpenListener() {
+            @Override
+            public void onOpened() {
+                camera.startStreaming(800, 448, OpenCvCameraRotation.UPRIGHT);
+            }
+
+            @Override
+            public void onError(int errorCode) {
+
+            }
+        });
+
         // Initialize the robot
         RobotCentricMechanumDrive drive = new RobotCentricMechanumDrive(hardwareMap, DcMotorSimple.Direction.REVERSE);
         drive.setTicks(TICKS_PER_INCH, TICKS_PER_INCH);
+        // whereas 3231 wants
+        // RobotCentricMechanumDrive drive = new RobotCentricMechanumDrive(hardwareMap, DcMotorSimple.Direction.REVERSE);
 
         // Initialize Operator Members and Motors
         // Grabber grabber = new Grabber(hardwareMap.servo.get(RobotConstants.GRABBER), Servo.Direction.REVERSE);
         // LinearSlider slider = new LinearSlider(hardwareMap.dcMotor.get(RobotConstants.SLIDE), DcMotorSimple.Direction.REVERSE);
+        // TODO: linearSlide.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
 
-        // Initialize camera
         telemetry.setMsTransmissionInterval(50);
-        AprilTagsCamera camera = new AprilTagsCamera();
 
-        AprilTagDetection tagOfInterest = null;
         while (!isStarted() && !isStopRequested()) {
-            tagOfInterest = camera.detectTag(telemetry);
+            ArrayList<AprilTagDetection> currentDetections = aprilTagDetectionPipeline.getLatestDetections();
+
+            if (currentDetections.size() != 0) {
+                boolean tagFound = false;
+                for (AprilTagDetection tag : currentDetections) {
+                    tagOfInterest = tag;
+                    tagFound = true;
+                    break;
+                }
+
+                if (tagFound) {
+                    telemetry.addLine("Tag of interest is in sight!\n\nLocation data:");
+                    tagToTelemetry(tagOfInterest);
+                } else {
+                    telemetry.addLine("Don't see tag of interest :(");
+                    if (tagOfInterest == null) {
+                        telemetry.addLine("(The tag has never been seen)");
+                    } else {
+                        telemetry.addLine("\nBut we HAVE seen the tag before; last seen at:");
+                        tagToTelemetry(tagOfInterest);
+                    }
+                }
+
+            } else {
+                telemetry.addLine("Don't see tag of interest :(");
+                if (tagOfInterest == null) {
+                    telemetry.addLine("(The tag has never been seen)");
+                } else {
+                    telemetry.addLine("\nBut we HAVE seen the tag before; last seen at:");
+                    tagToTelemetry(tagOfInterest);
+                }
+
+            }
+
+            telemetry.update();
+            sleep(20);
         }
 
         waitForStart();
-        camera.updateTelemetry(telemetry);
 
-        // Retrieve the position
-        int position = camera.getPosition(tagOfInterest);
-        telemetry.addLine("We are heading towards position #" + position + "!");
-        switch (position) {
-            case 1:
-                // Position #1 (Left)
+        /*
+         * The START command just came in: now work off the latest snapshot acquired
+         * during the init loop.
+         */
+
+        /* Update the telemetry */
+        if (tagOfInterest != null) {
+            telemetry.addLine("Tag snapshot:\n");
+            tagToTelemetry(tagOfInterest);
+            telemetry.update();
+        } else {
+            telemetry.addLine("No tag snapshot available, it was never sighted during the init loop :(");
+            telemetry.update();
+        }
+
+        /* Actually do something useful */
+        if (tagOfInterest == null) {
+            /*
+             * Insert your autonomous code here, presumably running some default configuration
+             * since the tag was never sighted during INIT
+             */
+            // Move forward to Position #2 for the 1/3 chance it works.
+            telemetry.addLine("Running Autonomous without tags");
+            telemetry.update();
+            drive.moveLinear(FORWARD, 16, 0.25, telemetry);
+        } else {
+            /*
+             * Insert your autonomous code here, probably using the tag pose to decide your configuration.
+             */
+
+            // e.g.
+            if (tagOfInterest.id == LEFT) {
+                // Drive to Position #1 (Left) w/ Encoders
+            // TODO:  This is the version for 2939.
                 drive.moveLinear(FORWARD, 24, 0.25, telemetry);
                 drive.strafe(LEFT_STRAFE, 26, 0.25, telemetry);
-                break;
-            case 2:
-                // Position #2 (Middle)
+            } else if (tagOfInterest.id == MIDDLE) {
+                // Drive to Position #2 (Middle) w/ Encoders
                 drive.moveLinear(FORWARD, 24, 0.25, telemetry);
-                break;
-            case 3:
-                // Position #3 (Right)
+            } else if (tagOfInterest.id == RIGHT) {
+                // Drive to Position #3 (Right) w/ Encoders
                 drive.moveLinear(FORWARD, 24, 0.25, telemetry);
                 drive.strafe(RIGHT_STRAFE, 26, 0.25, telemetry);
-                break;
-            default:
-                // Tag is not found
-                telemetry.addLine("Tag was not found.");
+            } else {
+                // Somehow we got another tag.. Drive to Position #2 (Middle) w/ one third chance.
+                drive.moveLinear(FORWARD, 24, 0.25, telemetry);
+            }
         }
-        telemetry.update();
+    }
+
+    void tagToTelemetry(AprilTagDetection detection) {
+        telemetry.addLine(String.format("\nDetected tag ID=%d", detection.id));
+        telemetry.addLine(String.format("Translation X: %.2f feet", detection.pose.x * FEET_PER_METER));
+        telemetry.addLine(String.format("Translation Y: %.2f feet", detection.pose.y * FEET_PER_METER));
+        telemetry.addLine(String.format("Translation Z: %.2f feet", detection.pose.z * FEET_PER_METER));
+        telemetry.addLine(String.format("Rotation Yaw: %.2f degrees", Math.toDegrees(detection.pose.yaw)));
+        telemetry.addLine(String.format("Rotation Pitch: %.2f degrees", Math.toDegrees(detection.pose.pitch)));
+        telemetry.addLine(String.format("Rotation Roll: %.2f degrees", Math.toDegrees(detection.pose.roll)));
     }
 }
