--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/lib/AprilTagsCamera.java
@@ -1,125 +1,0 @@
-package org.firstinspires.ftc.teamcode.lib;
-
-import static org.firstinspires.ftc.robotcore.external.BlocksOpModeCompanion.hardwareMap;
-
-import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
-import org.firstinspires.ftc.teamcode.Autonomous.AprilTagDetectionPipeline;
-import org.openftc.apriltag.AprilTagDetection;
-import org.openftc.easyopencv.OpenCvCamera;
-import org.openftc.easyopencv.OpenCvCameraFactory;
-import org.openftc.easyopencv.OpenCvCameraRotation;
-
-import java.util.ArrayList;
-
-public class AprilTagsCamera {
-    OpenCvCamera camera;
-    AprilTagDetectionPipeline aprilTagDetectionPipeline;
-
-    static final double FEET_PER_METER = 3.28084;
-
-    // NOTE: this calibration is for the C920 webcam at 800x448.
-    double fx = 578.272;
-    double fy = 578.272;
-    double cx = 402.145;
-    double cy = 221.506;
-
-    double tagsize = 0.166;
-
-    int LEFT = 11;
-    int MIDDLE = 12;
-    int RIGHT = 13;
-
-    AprilTagDetection tagOfInterest = null;
-
-    public AprilTagsCamera() {
-        int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier("cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
-        camera = OpenCvCameraFactory.getInstance().createWebcam(hardwareMap.get(WebcamName.class, "Webcam 1"), cameraMonitorViewId);
-        aprilTagDetectionPipeline = new AprilTagDetectionPipeline(tagsize, fx, fy, cx, cy);
-
-        camera.setPipeline(aprilTagDetectionPipeline);
-        camera.openCameraDeviceAsync(new OpenCvCamera.AsyncCameraOpenListener() {
-            @Override
-            public void onOpened() {
-                camera.startStreaming(800, 448, OpenCvCameraRotation.UPRIGHT);
-            }
-
-            @Override
-            public void onError(int errorCode) {
-
-            }
-        });
-    }
-
-    public void tagToTelemetry(AprilTagDetection detection, Telemetry telemetry) {
-        telemetry.addLine(String.format("\nDetected tag ID=%d", detection.id));
-        telemetry.addLine(String.format("Translation X: %.2f feet", detection.pose.x * FEET_PER_METER));
-        telemetry.addLine(String.format("Translation Y: %.2f feet", detection.pose.y * FEET_PER_METER));
-        telemetry.addLine(String.format("Translation Z: %.2f feet", detection.pose.z * FEET_PER_METER));
-        telemetry.addLine(String.format("Rotation Yaw: %.2f degrees", Math.toDegrees(detection.pose.yaw)));
-        telemetry.addLine(String.format("Rotation Pitch: %.2f degrees", Math.toDegrees(detection.pose.pitch)));
-        telemetry.addLine(String.format("Rotation Roll: %.2f degrees", Math.toDegrees(detection.pose.roll)));
-    }
-
-    public AprilTagDetection detectTag(Telemetry telemetry) throws InterruptedException {
-        ArrayList<AprilTagDetection> currentDetections = aprilTagDetectionPipeline.getLatestDetections();
-        if (currentDetections.size() != 0) {
-            boolean tagFound = false;
-            for (AprilTagDetection tag : currentDetections) {
-                tagOfInterest = tag;
-                tagFound = true;
-                break;
-            }
-
-            if (tagFound) {
-                telemetry.addLine("Tag of interest is in sight!\n\nLocation data:");
-                tagToTelemetry(tagOfInterest, telemetry);
-            } else {
-                telemetry.addLine("Don't see tag of interest :(");
-                if (tagOfInterest == null) {
-                    telemetry.addLine("(The tag has never been seen)");
-                } else {
-                    telemetry.addLine("\nBut we HAVE seen the tag before; last seen at:");
-                    tagToTelemetry(tagOfInterest, telemetry);
-                }
-            }
-
-        } else {
-            telemetry.addLine("Don't see tag of interest :(");
-            if (tagOfInterest == null) {
-                telemetry.addLine("(The tag has never been seen)");
-            } else {
-                telemetry.addLine("\nBut we HAVE seen the tag before; last seen at:");
-                tagToTelemetry(tagOfInterest, telemetry);
-            }
-
-        }
-        telemetry.update();
-        Thread.sleep(20);
-        return tagOfInterest;
-    }
-
-    public void updateTelemetry(Telemetry telemetry) {
-        /* Update the telemetry */
-        if (tagOfInterest != null) {
-            telemetry.addLine("Tag snapshot:\n");
-            tagToTelemetry(tagOfInterest, telemetry);
-            telemetry.update();
-        } else {
-            telemetry.addLine("No tag snapshot available, it was never sighted during the init loop :(");
-            telemetry.update();
-        }
-    }
-
-    public int getPosition(AprilTagDetection tag) {
-        if (tag.id == LEFT) {
-            return 1;
-        } else if (tag.id == MIDDLE) {
-            return 2;
-        } else if (tag.id == RIGHT) {
-            return 3;
-        }
-
-        return -999;
-    }
-}
